requirements.txt
streamlit
sounddevice
soundfile
numpy
SpeechRecognition
requests
pyttsx3
gtts
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langdetect
